requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pytesseract==0.3.10
Werkzeug==3.0.3
